Fix WarningAction parameter in MicrosoftPowerAutomateDesktop (#46)
* Add support for Microsoft Power Automate Desktop

Introduced Get-MicrosoftPowerAutomateDesktop.ps1 to retrieve the latest version and download URL for Microsoft Power Automate Desktop.

* Update Evergreen module version requirement

Bumped the required Evergreen module version from 2510.2799.0 to 2510.2802.0 in Get-MicrosoftPowerAutomateDesktop.ps1 to ensure compatibility with the latest features or fixes.

* Remove suppress message attribute and update warning action

Removed the SuppressMessage attribute for singular noun usage and changed the WarningAction parameter from 'Ignore' to 'SilentlyContinue' in the Get-MicrosoftPowerAutomateDesktop.ps1 script.